Add files via upload
The project is a database for a makeup store that contains 8 tables (Products, Brands, Categories, Subcategories, Customers, Orders, Order Items, and Product Reviews). The tables are linked as shown in the image included with the project.  
There is a Windows Form .NET Framework linked to the database, so when the admin logs in, they can see a report for all customers, products with the highest reviews, and best-selling products.  
To log in to this Windows Form, there is a login table in the database that includes a **Username** and **Password** for the admin.  

My project file includes the database backup. It also includes an SQL file so you can open SQL Server and execute the commands to create the database and tables (Don’t forget to change the username and password in the login table. If you don’t change it, they will be **Username: Joudi** and **Password: J12345**). It also includes the Windows Form where you can log in and see the report (Don’t forget to change the server name in **Form1.cs** and **Raporlamacs.cs**). It also includes a PNG for the ER diagram of the database and a Word file with the business rules for the database.